Enforce module boundary types
This rule protects against typescript inferring an overly specific
subtype of the desired module boundary type, or the developer accidentally
changing the desired module boundary type when refactoring.

Example: Given the class structure, A, B extends A, and C extends A. If
I have a module with a function that returns a B instance, typescript
will infer type B, when in reality I may want the module to only expose
that it return a type A. This would allow me to refactor to returning a
C without breaking anything because it forces me to define the boundary
of my module from the start.